pyproject.toml
[project]
name = "hack"
version = "0.1.0"
description = "DIS2026X1 robot-agent hackathon scaffolding"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.12",
    "rich>=13.7",
    "pydantic>=2.7",
    "pyyaml>=6.0",
    "httpx>=0.27",
    "opencv-python>=4.10",
    "numpy>=1.26",
    "pillow>=10.3",
    "sounddevice>=0.4.7",
    "soundfile>=0.12",
    "fastapi>=0.115",
    "uvicorn>=0.30",
    "jinja2>=3.1",
    "python-multipart>=0.0.9",
    "psutil>=5.9",
]

[project.optional-dependencies]
audio = [
    "faster-whisper>=1.0",
    "silero-vad>=5.1",
    "piper-tts>=1.2; platform_system != 'Darwin'",
]
llm = [
    "ollama>=0.3",
    "openai>=1.40",
    "anthropic>=0.34",
]
# Optional robotics drivers. Install on the laptop where a physical robot is attached.
# LeRobot wheels are big (PyTorch etc.) — keep out of the default install.
robot = [
    "lerobot>=0.1; python_version >= '3.11'",
]
dev = [
    "pytest>=8.0",
    "ruff>=0.6",
    "pyright>=1.1",
]

[project.scripts]
hack = "hack.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/hack"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pyright]
include = ["src", "tests"]
pythonVersion = "3.11"
